rename CartoMapV3Table to CartoMapV3
