+++ lib/Traits/TraitsClasses/CartoMapV3CatalogItemTraits.ts
